+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.2
   - Features Added
+    - First loading speed improvement
     - Support Multi-call Generic Scheme
     - Use subgraph v40_0 and arc.js 0.2.76
     - Added xGEN / GEN bridge